requirements.txt
Flask>=2.0
python-dotenv>=0.21.0
requests>=2.0
google-api-python-client>=2.0
google-auth>=2.0
